+++ 0x0A-argc_argv/main.h
@@ -4,6 +4,6 @@
 
 #include <stdio.h>
 int _putchar(char c);
-int atoi(char *str);
+float atoi(char *str);
 
 #endif /* I'm Solo -__= */
